More cypress (#120)
* fix: gitpod

* fix: approve test

* fix: refactor cypress

* fix: gitpod config

* fix: gitpod config

* fix: gitpod + remove console log

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: approve test

* fix: refactor cypress

* fix: gitpod config

* fix: gitpod config

* fix: gitpod + remove console log

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: gitpod

* fix: minor

* fix: minor
