+++ .gitpod.yml
@@ -2,22 +2,26 @@ image:
   file: .gitpod.Dockerfile
 
 tasks:
+  - init: docker-compose down
+    command: docker-compose up
+
   - init: yarn install --frozen-lockfile
     command: |
       export $(cat .env | grep -v "#" | xargs)
-      MONGODB_HOST=localhost ./apps/artemis-web/scripts/wait-for -h ${MONGODB_HOST} -p ${MONGODB_PORT} -t 0
+      MONGODB_HOST=localhost ./apps/artemis-web/scripts/wait-for -h localhost -p ${MONGODB_PORT} -t 0
+      yarn build --prod
       MONGODB_HOST=localhost node apps/artemis-web/scripts/query.js
-      MONGODB_HOST=localhost LDAP_HOST=localhost yarn run start
-  - init: |
-      mkdir -p /workspace/data && mongod --dbpath /workspace/data &
-      sleep 5
-      if [ -f .env ]; then export $(cat .env | sed 's/#.*//g' | xargs); fi
-      mongo admin --eval "db.createUser({user:'"$MONGODB_USER"',pwd:'"$MONGODB_PASS"',roles:[{role:'root',db:'admin'}]})"
-      killall mongod
-    command: mongod --auth --dbpath /workspace/data
+      npx cypress install
+      MONGODB_HOST=localhost LDAP_HOST=localhost yarn start --prod
 ports:
   - port: 4200
     onOpen: open-browser
+  - port: 5900
+    onOpen: ignore
+  - port: 6080
+    onOpen: open-browser
+  - port: 10000
+    onOpen: ignore
 github:
   prebuilds:
     # enable for the master/default branch (defaults to true)
